Moved Tool(s)Stats to core
This commit moves several user-related types to core, to resolve a
cyclic dependency.

See: gh-1024
