+++ embabel-agent-api/src/main/kotlin/com/embabel/agent/spi/support/AgenticEventListenerToolsStats.kt
@@ -15,8 +15,8 @@
  */
 package com.embabel.agent.spi.support
 
-import com.embabel.agent.api.common.ToolStats
-import com.embabel.agent.api.common.ToolsStats
+import com.embabel.agent.core.ToolStats
+import com.embabel.agent.core.ToolsStats
 import com.embabel.agent.api.event.AgentProcessEvent
 import com.embabel.agent.api.event.AgenticEventListener
 import com.embabel.agent.api.event.ToolCallResponseEvent
