+++ embabel-agent-api/src/test/kotlin/com/embabel/agent/core/ToolsStatsTest.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.embabel.agent.api.common
+package com.embabel.agent.core
 
 import io.mockk.every
 import io.mockk.mockk
