+++ embabel-agent-api/src/test/kotlin/com/embabel/agent/spi/support/springai/streaming/StreamingChatClientOperationsGuardRailTest.kt
@@ -18,7 +18,7 @@
 package com.embabel.agent.spi.support.springai.streaming
 
 import com.embabel.agent.api.common.InteractionId
-import com.embabel.agent.api.common.ToolsStats
+import com.embabel.agent.core.ToolsStats
 import com.embabel.agent.api.event.LlmRequestEvent
 import com.embabel.agent.api.event.observation.InternalObservabilityApi
 import com.embabel.agent.api.validation.guardrails.GuardRailViolationException
